Create a new server.js file
